+++ lib/liquex/tag/object_tag.ex
@@ -105,11 +105,8 @@ defmodule Liquex.Tag.ObjectTag do
   end
 
   def render([argument, filters: filters], %Context{} = context) do
-    {result, context} =
-      argument
-      |> List.wrap()
-      |> Liquex.Argument.eval(context)
-      |> Render.apply_filters(filters, context)
+    {value, context} = Liquex.Argument.eval(List.wrap(argument), context)
+    {result, context} = Render.apply_filters(value, filters, context)
 
     {Render.to_output_string(result), context}
   end